+++ api/application/svg_builder.py
@@ -1,0 +1,4 @@
+def render(data):
+    return '''<svg width="50" height="50">
+        <circle cx="25" cy="25" r="20"/>
+    </svg>'''
